Entity Framework Push

+++ Entities/Concrete/Car.cs
@@ -10,7 +10,7 @@ namespace Entities.Concrete
         public int Id { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
-        public long ModelYear { get; set; }
+        public short ModelYear { get; set; }
         public int DailyPrice { get; set; }
         public string Description { get; set; }
     }
